+++ NEWS.md
@@ -17,6 +17,7 @@ Some examples:
 - Thanks to David Watson, parallel computing is now supported. The user needs to set up the parallel backend before calling `kernelshap()`, e.g., using the "doFuture" package, and then set `parallel = TRUE`. Especially on Windows, sometimes not all global variables or packages are loaded in the parallel instances. These can be specified by `parallel_args`, a list of arguments passed to `foreach()`.
 - Even without parallel computing, `kernelshap()` has become much faster.
 - For $2 \le p \le 5$ features, the algorithm now returns exact Kernel SHAP values with respect to the given background data. (For $p = 1$, exact *Shapley values* are returned.)
+- Direct handling of "tidymodels" models.
 
 ## User visible changes